adds `UpdatableRecordTable` type
This interface is very simple, it permits to initialize a table for some
distance, to update the value for each distance, using the `append`
method (No verification that the record is actually better than the
previous one will be done, it's just a simple display table), and then
when everything is settle down, just seal the display.

- Adds also an implementation for the console, using the `readline` library.
- And a simple stream to be able to write in the table using object
stream of records.

WIP: as it is kind of difficult to unit test mutated console outputs,
some manual tests have been made, and are currently launchable using:
```
ts-node src/output/UpdatableRecordTableStream.ts
```
This test will be removed as soon as another real class will use the
stream.