extracts record find algo into a transform stream
Introduces `FindRecordsStream`.
This stream is then used by the original `Records` method, which will
just write all positions, and gather records from the stream.

So currently output of the `FindRecordsStream` is ignored, and we gather
state from the trackers.
But in the near future, `extractBests` should be removed, so do
`Records`, we might do that as soon as some more plumbing using pipes
will be done.