allows multiple gpx files
- introduces a new stream to aggregate the records, every file is
read, parsed, and record retrieved in a strem pipeline. And all of them
are writing to the same aggregator stream, which keep only the best
records for each distance.
This stream will also notify the table when some new records are stored.

wip: remove dead code (`Records`) and modify test accordingly (maybe
just test the find records stream, and the aggregator).