introduces stream merger to merge inputs stream
Like this it's way more easily to deal with multiple inputs piped to the
same output.
It deals with the max listeners, and the 'end' event.
So it avoid verbosity, and permit to write the pipe to the output only
once.
So it's also nicer to read in `index.ts`.