introduces `peek`, to listen on stream
It permit to consume (synchronously) chunks at some point of the pipelines without
altering it (in case it does not throw any error!).

In the mean time, and to test this new functionnality, table is from now
on able to have a status bar and to display the number of parsed
positions. So this new `peek` method is used to notify the table of all
the parsed positions.

Also add a time calculation at the end of the main process, adds some
emoji, and also remove emojis for records. Too much emojis kill emojis?!